:truck: Move `CampaignReport` to `/src/prescription/campaign/`
