+++ api/tests/tooling/domain-builder/factory/build-campaign-report.js
@@ -1,5 +1,5 @@
+import { CampaignReport } from '../../../../src/prescription/campaign/domain/read-models/CampaignReport.js';
 import { CampaignExternalIdTypes, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
-import { CampaignReport } from '../../../../src/shared/domain/read-models/CampaignReport.js';
 
 const buildCampaignReport = function ({
   id = 1,
